+++ src/Oro/Bundle/ScopeBundle/Tests/Unit/Manager/ScopeManagerTest.php
@@ -2,16 +2,18 @@
 
 namespace Oro\Bundle\ScopeBundle\Tests\Unit\Manager;
 
-use Oro\Bundle\EntityBundle\Provider\EntityFieldProvider;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Oro\Bundle\ScopeBundle\Entity\Repository\ScopeRepository;
 use Oro\Bundle\ScopeBundle\Entity\Scope;
-use Oro\Bundle\ScopeBundle\Manager\ScopeCriteriaProviderInterface;
 use Oro\Bundle\ScopeBundle\Manager\ScopeEntityStorage;
 use Oro\Bundle\ScopeBundle\Manager\ScopeManager;
 use Oro\Bundle\ScopeBundle\Model\ScopeCriteria;
+use Oro\Bundle\ScopeBundle\Tests\Unit\Stub\StubContext;
 use Oro\Bundle\ScopeBundle\Tests\Unit\Stub\StubScope;
 use Oro\Bundle\ScopeBundle\Tests\Unit\Stub\StubScopeCriteriaProvider;
-use Oro\Component\TestUtils\Mocks\ServiceLink;
+use Psr\Container\ContainerInterface;
+use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
+use Symfony\Component\PropertyAccess\PropertyAccessor;
 
 /**
  * @SuppressWarnings(PHPMD.TooManyMethods)
@@ -21,135 +23,298 @@ use Oro\Component\TestUtils\Mocks\ServiceLink;
  */
 class ScopeManagerTest extends \PHPUnit\Framework\TestCase
 {
-    /**
-     * @var ScopeManager|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $manager;
+    /** @var ScopeEntityStorage|\PHPUnit\Framework\MockObject\MockObject */
+    private $entityStorage;
 
-    /**
-     * @var ScopeEntityStorage|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $entityStorage;
+    /** @var ClassMetadata|\PHPUnit\Framework\MockObject\MockObject */
+    private $scopeClassMetadata;
 
-    /**
-     * @var EntityFieldProvider|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $entityFieldProvider;
+    /** @var ScopeRepository|\PHPUnit\Framework\MockObject\MockObject */
+    private $scopeRepository;
 
-    public function setUp()
+    protected function setUp()
     {
         $this->entityStorage = $this->createMock(ScopeEntityStorage::class);
-        $this->entityFieldProvider = $this->createMock(EntityFieldProvider::class);
+        $this->scopeClassMetadata = $this->createMock(ClassMetadata::class);
+        $this->scopeRepository = $this->createMock(ScopeRepository::class);
 
-        $serviceLink = new ServiceLink($this->entityFieldProvider);
-        $this->manager = new ScopeManager($this->entityStorage, $serviceLink);
+        $this->entityStorage->expects($this->any())
+            ->method('getClassMetadata')
+            ->willReturn($this->scopeClassMetadata);
+        $this->entityStorage->expects($this->any())
+            ->method('getRepository')
+            ->willReturn($this->scopeRepository);
     }
 
-    public function tearDown()
+    /**
+     * @param array $providers
+     *
+     * @return ScopeManager
+     */
+    private function getScopeManager(array $providers = []): ScopeManager
     {
-        unset($this->manager, $this->registry, $this->entityFieldProvider);
+        $serviceMap = [];
+        $providerIds = [];
+        foreach ($providers as $scopeType => $services) {
+            $serviceIds = [];
+            foreach ($services as $key => $service) {
+                $serviceId = sprintf('%s_%s', $scopeType, $key);
+                $serviceIds[] = $serviceId;
+                $serviceMap[$serviceId] = $service;
+            }
+            $providerIds[$scopeType] = $serviceIds;
+        }
+
+        $container = $this->createMock(ContainerInterface::class);
+        $container->expects($this->any())
+            ->method('get')
+            ->willReturnCallback(function ($serviceId) use ($serviceMap) {
+                if (!isset($serviceMap[$serviceId])) {
+                    throw new ServiceNotFoundException($serviceId);
+                }
+
+                return $serviceMap[$serviceId];
+            });
+
+        return new ScopeManager(
+            $container,
+            $providerIds,
+            $this->entityStorage,
+            new PropertyAccessor()
+        );
     }
 
     public function testFindDefaultScope()
     {
-        $this->entityFieldProvider->expects($this->once())
-            ->method('getRelations')
-            ->with(Scope::class, false, true, false)
-            ->willReturn([['name' => 'relation']]);
-        $expectedCriteria = new ScopeCriteria(['relation' => null], [['name' => 'relation']]);
-        $repository = $this->getMockBuilder(ScopeRepository::class)->disableOriginalConstructor()->getMock();
+        $expectedCriteria = new ScopeCriteria(['relation' => null], $this->scopeClassMetadata);
         $scope = new Scope();
-        $repository->method('findOneByCriteria')
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['relation']);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
             ->with($expectedCriteria)
             ->willReturn($scope);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->assertSame($scope, $this->manager->findDefaultScope());
+        $manager = $this->getScopeManager();
+        $this->assertSame($scope, $manager->findDefaultScope());
     }
 
     public function testGetCriteriaByScope()
     {
-        $this->manager->addProvider('test', new StubScopeCriteriaProvider());
         $scope = new StubScope();
         $scope->setScopeField('expected_value');
-        $this->entityFieldProvider->method('getRelations')->willReturn([]);
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([]);
+
+        $provider = new StubScopeCriteriaProvider('scopeField', new \stdClass(), \stdClass::class);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
         $this->assertSame(
-            [StubScopeCriteriaProvider::STUB_FIELD => 'expected_value'],
-            $this->manager->getCriteriaByScope($scope, 'test')->toArray()
+            [$provider->getCriteriaField() => 'expected_value'],
+            $manager->getCriteriaByScope($scope, 'testScope')->toArray()
         );
     }
 
     public function testFind()
     {
         $scope = new Scope();
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaForCurrentScope')->willReturn(['fieldName' => 1]);
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $scopeCriteria = new ScopeCriteria(['fieldName' => 1, 'fieldName2' => null], $fieldsInfo);
-        $repository = $this->getMockBuilder(ScopeRepository::class)->disableOriginalConstructor()->getMock();
-        $repository->method('findOneByCriteria')->with($scopeCriteria)->willReturn($scope);
-
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->entityFieldProvider->method('getRelations')->willReturn(
-            [
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => $fieldValue, 'fieldName2' => null],
+            $this->scopeClassMetadata
         );
 
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->find('testScope');
-        $this->assertEquals($scope, $actualScope);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', $fieldValue, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn($scope);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope'));
+    }
+
+    public function testFindWithArrayContext()
+    {
+        $scope = new Scope();
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['field' => $fieldValue, 'field2' => null],
+            $this->scopeClassMetadata
+        );
+        $context = ['field' => $fieldValue];
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['field', 'field2']);
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn($scope);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope', $context));
+    }
+
+    public function testFindWithObjectContext()
+    {
+        $scope = new Scope();
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['field' => $fieldValue, 'field2' => null],
+            $this->scopeClassMetadata
+        );
+        $context = new StubContext();
+        $context->setField($fieldValue);
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['field', 'field2']);
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn($scope);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope', $context));
+    }
+
+    public function testFindWithIsNotNullValueInContext()
+    {
+        $scope = new Scope();
+        $fieldValue = ScopeCriteria::IS_NOT_NULL;
+        $scopeCriteria = new ScopeCriteria(
+            ['field' => $fieldValue, 'field2' => null],
+            $this->scopeClassMetadata
+        );
+        $context = ['field' => $fieldValue];
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['field', 'field2']);
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn($scope);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope', $context));
+    }
+
+    public function testFindWithArrayValueInContext()
+    {
+        $scope = new Scope();
+        $fieldValue = [1, 2, 3];
+        $scopeCriteria = new ScopeCriteria(
+            ['field' => $fieldValue, 'field2' => null],
+            $this->scopeClassMetadata
+        );
+        $context = ['field' => $fieldValue];
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['field', 'field2']);
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn($scope);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope', $context));
+    }
+
+    // @codingStandardsIgnoreStart
+    /**
+     * @expectedException \Oro\Bundle\ScopeBundle\Exception\NotSupportedCriteriaValueException
+     * @expectedExceptionMessage The type string is not supported for context[field]. Expected stdClass, null, array or "IS_NOT_NULL".
+     */
+    // @codingStandardsIgnoreEnd
+    public function testFindWithInvalidScalarValueInContext()
+    {
+        $context = ['field' => 'test'];
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->never())
+            ->method('findOneByCriteria');
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $manager->find('testScope', $context);
+    }
+
+    // @codingStandardsIgnoreStart
+    /**
+     * @expectedException \Oro\Bundle\ScopeBundle\Exception\NotSupportedCriteriaValueException
+     * @expectedExceptionMessage The type Oro\Bundle\ScopeBundle\Tests\Unit\Stub\StubScope is not supported for context[field]. Expected stdClass, null, array or "IS_NOT_NULL".
+     */
+    // @codingStandardsIgnoreEnd
+    public function testFindWithInvalidObjectValueInContext()
+    {
+        $context = ['field' => new StubScope()];
+
+        $provider = new StubScopeCriteriaProvider('field', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->never())
+            ->method('findOneByCriteria');
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $manager->find('testScope', $context);
     }
 
     public function testFindScheduled()
     {
         $scope = new Scope();
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->expects($this->any())
-            ->method('getCriteriaForCurrentScope')
-            ->willReturn(['fieldName' => 1]);
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => $fieldValue, 'fieldName2' => null],
+            $this->scopeClassMetadata
+        );
 
-        $this->entityFieldProvider->expects($this->any())
-            ->method('getRelations')
-            ->willReturn([
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
 
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $scopeCriteria = new ScopeCriteria(['fieldName' => 1, 'fieldName2' => null], $fieldsInfo);
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
+        $provider = new StubScopeCriteriaProvider('fieldName', $fieldValue, \stdClass::class);
 
-        $repository->expects($this->once())
+        $this->scopeRepository->expects($this->once())
             ->method('findOneByCriteria')
             ->with($scopeCriteria);
-
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
 
         $this->entityStorage->expects($this->once())
             ->method('getScheduledForInsertByCriteria')
             ->with($scopeCriteria)
             ->willReturn($scope);
 
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->find('testScope');
-        $this->assertEquals($scope, $actualScope);
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->find('testScope'));
     }
 
     public function testCreateScopeByCriteriaWithFlush()
     {
-        $scopeCriteria = new ScopeCriteria([], []);
+        $scopeCriteria = new ScopeCriteria([], $this->scopeClassMetadata);
+
         $this->entityStorage->expects($this->once())
             ->method('getScheduledForInsertByCriteria')
             ->with($scopeCriteria)
@@ -160,12 +325,14 @@ class ScopeManagerTest extends \PHPUnit\Framework\TestCase
         $this->entityStorage->expects($this->once())
             ->method('flush');
 
-        $this->assertInstanceOf(Scope::class, $this->manager->createScopeByCriteria($scopeCriteria));
+        $manager = $this->getScopeManager();
+        $this->assertInstanceOf(Scope::class, $manager->createScopeByCriteria($scopeCriteria));
     }
 
     public function testCreateScopeByCriteriaWithoutFlush()
     {
-        $scopeCriteria = new ScopeCriteria([], []);
+        $scopeCriteria = new ScopeCriteria([], $this->scopeClassMetadata);
+
         $this->entityStorage->expects($this->once())
             ->method('getScheduledForInsertByCriteria')
             ->with($scopeCriteria)
@@ -176,13 +343,15 @@ class ScopeManagerTest extends \PHPUnit\Framework\TestCase
         $this->entityStorage->expects($this->never())
             ->method('flush');
 
-        $this->assertInstanceOf(Scope::class, $this->manager->createScopeByCriteria($scopeCriteria, false));
+        $manager = $this->getScopeManager();
+        $this->assertInstanceOf(Scope::class, $manager->createScopeByCriteria($scopeCriteria, false));
     }
 
     public function testCreateScopeByCriteriaScheduled()
     {
         $scope = new Scope();
-        $scopeCriteria = new ScopeCriteria([], []);
+        $scopeCriteria = new ScopeCriteria([], $this->scopeClassMetadata);
+
         $this->entityStorage->expects($this->once())
             ->method('getScheduledForInsertByCriteria')
             ->with($scopeCriteria)
@@ -192,258 +361,210 @@ class ScopeManagerTest extends \PHPUnit\Framework\TestCase
         $this->entityStorage->expects($this->never())
             ->method('flush');
 
-        $this->assertEquals($scope, $this->manager->createScopeByCriteria($scopeCriteria));
+        $manager = $this->getScopeManager();
+        $this->assertEquals($scope, $manager->createScopeByCriteria($scopeCriteria));
     }
 
     public function testFindBy()
     {
         $scope = new Scope();
-        $scopeCriteria = new ScopeCriteria(
-            [StubScopeCriteriaProvider::STUB_FIELD => StubScopeCriteriaProvider::STUB_VALUE],
-            [['name' => StubScopeCriteriaProvider::STUB_FIELD]]
-        );
-        $repository = $this->getMockBuilder(ScopeRepository::class)->disableOriginalConstructor()->getMock();
-        $repository->method('findByCriteria')->with($scopeCriteria)->willReturn([$scope]);
+        $criteriaField = 'scopeField';
+        $criteriaValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria([$criteriaField => $criteriaValue], $this->scopeClassMetadata);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([$criteriaField]);
 
-        $this->entityFieldProvider->method('getRelations')->willReturn(
-            [
-                ['name' => StubScopeCriteriaProvider::STUB_FIELD],
-            ]
-        );
+        $provider = new StubScopeCriteriaProvider($criteriaField, $criteriaValue, \stdClass::class);
 
-        $this->manager->addProvider('testScope', new StubScopeCriteriaProvider());
-        $this->assertEquals([$scope], $this->manager->findBy('testScope'));
+        $this->scopeRepository->expects($this->once())
+            ->method('findByCriteria')
+            ->with($scopeCriteria)
+            ->willReturn([$scope]);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals([$scope], $manager->findBy('testScope'));
     }
 
     public function testFindRelatedScopes()
     {
-        $this->entityFieldProvider->method('getRelations')->willReturn(
-            [
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]
-        );
-        /** @var ScopeCriteriaProviderInterface|\PHPUnit\Framework\MockObject\MockObject $provider */
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaByContext')
-            ->willReturn([]);
-        $provider->method('getCriteriaField')
-            ->willReturn('fieldName');
-        $this->manager->addProvider('test', $provider);
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
         $scopes = [new Scope()];
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $repository->expects($this->once())
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => ScopeCriteria::IS_NOT_NULL, 'fieldName2' => null],
+            $this->scopeClassMetadata
+        );
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
             ->method('findByCriteria')
-            ->with(new ScopeCriteria(['fieldName' => ScopeCriteria::IS_NOT_NULL, 'fieldName2' => null], $fieldsInfo))
+            ->with($scopeCriteria)
             ->willReturn($scopes);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->assertSame($scopes, $this->manager->findRelatedScopes('test'));
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertSame($scopes, $manager->findRelatedScopes('testScope'));
     }
 
     public function testFindRelatedScopeIds()
     {
-        $this->entityFieldProvider->method('getRelations')->willReturn(
-            [
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]
-        );
-        /** @var ScopeCriteriaProviderInterface|\PHPUnit\Framework\MockObject\MockObject $provider */
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaByContext')
-            ->willReturn([]);
-        $provider->method('getCriteriaField')
-            ->willReturn('fieldName');
-        $this->manager->addProvider('test', $provider);
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
         $scopeIds = [1, 4];
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $repository->expects($this->once())
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => ScopeCriteria::IS_NOT_NULL, 'fieldName2' => null],
+            $this->scopeClassMetadata
+        );
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
             ->method('findIdentifiersByCriteria')
-            ->with(new ScopeCriteria(['fieldName' => ScopeCriteria::IS_NOT_NULL, 'fieldName2' => null], $fieldsInfo))
+            ->with($scopeCriteria)
             ->willReturn($scopeIds);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->assertSame($scopeIds, $this->manager->findRelatedScopeIds('test'));
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertSame($scopeIds, $manager->findRelatedScopeIds('testScope'));
     }
 
     public function testFindRelatedScopeIdsWithPriority()
     {
-        $this->entityFieldProvider->method('getRelations')->willReturn(
-            [
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]
-        );
-        /** @var ScopeCriteriaProviderInterface|\PHPUnit\Framework\MockObject\MockObject $provider */
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaByContext')
-            ->willReturn([]);
-        $provider->method('getCriteriaForCurrentScope')->willReturn(['fieldName' => 1]);
-        $this->manager->addProvider('test', $provider);
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
         $scopes = [new Scope()];
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $repository->expects($this->once())
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => $fieldValue, 'fieldName2' => null],
+            $this->scopeClassMetadata
+        );
+
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', $fieldValue, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
             ->method('findIdentifiersByCriteriaWithPriority')
-            ->with(new ScopeCriteria(['fieldName' => 1, 'fieldName2' => null], $fieldsInfo))
+            ->with($scopeCriteria)
             ->willReturn($scopes);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->assertSame($scopes, $this->manager->findRelatedScopeIdsWithPriority('test'));
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertSame($scopes, $manager->findRelatedScopeIdsWithPriority('testScope'));
     }
 
     public function testFindOrCreate()
     {
         $scope = new Scope();
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaForCurrentScope')->willReturn([]);
 
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $repository->method('findOneByCriteria')->willReturn(null);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([]);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-        $this->entityStorage->expects($this->once())
-            ->method('getRepository')
-            ->willReturn($repository);
+        $provider = new StubScopeCriteriaProvider('fieldName', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->willReturn(null);
+
         $this->entityStorage->expects($this->once())
             ->method('scheduleForInsert')
             ->with($scope, $this->isInstanceOf(ScopeCriteria::class));
         $this->entityStorage->expects($this->once())
             ->method('flush');
 
-        $this->entityFieldProvider->method('getRelations')->willReturn([]);
-
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->findOrCreate('testScope');
-        $this->assertEquals($scope, $actualScope);
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->findOrCreate('testScope'));
     }
 
     public function testFindOrCreateWithoutFlush()
     {
         $scope = new Scope();
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->expects($this->atLeastOnce())
-            ->method('getCriteriaForCurrentScope')->willReturn([]);
 
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $repository->expects($this->once())
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([]);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
             ->method('findOneByCriteria')
             ->willReturn(null);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
         $this->entityStorage->expects($this->once())
             ->method('scheduleForInsert')
             ->with($scope, $this->isInstanceOf(ScopeCriteria::class));
         $this->entityStorage->expects($this->never())
             ->method('flush');
 
-        $this->entityFieldProvider->expects($this->once())
-            ->method('getRelations')
-            ->willReturn([]);
-
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->findOrCreate('testScope', null, false);
-        $this->assertEquals($scope, $actualScope);
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->findOrCreate('testScope', null, false));
     }
 
     public function testFindOrCreateUsingContext()
     {
         $scope = new Scope();
         $context = ['scopeAttribute' => new \stdClass()];
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->method('getCriteriaByContext')->with($context)->willReturn([]);
 
-        $repository = $this->getMockBuilder(ScopeRepository::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $repository->method('findOneByCriteria')->willReturn(null);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([]);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
+        $provider = new StubScopeCriteriaProvider('fieldName', null, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
+            ->method('findOneByCriteria')
+            ->willReturn(null);
+
         $this->entityStorage->expects($this->once())
             ->method('scheduleForInsert')
             ->with($scope, $this->isInstanceOf(ScopeCriteria::class));
         $this->entityStorage->expects($this->once())
             ->method('flush');
 
-        $this->entityFieldProvider->method('getRelations')->willReturn([]);
-
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->findOrCreate('testScope', $context);
-        $this->assertEquals($scope, $actualScope);
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->findOrCreate('testScope', $context));
     }
 
     public function testGetScopeEntities()
     {
-        $this->manager->addProvider('scope_type', new StubScopeCriteriaProvider());
-        $expected = [
-            StubScopeCriteriaProvider::STUB_FIELD => StubScopeCriteriaProvider::STUB_CLASS
-        ];
+        $provider = new StubScopeCriteriaProvider('scopeField', null, \stdClass::class);
 
-        $this->assertEquals($expected, $this->manager->getScopeEntities('scope_type'));
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals(
+            [
+                $provider->getCriteriaField() => $provider->getCriteriaValueType()
+            ],
+            $manager->getScopeEntities('testScope')
+        );
     }
 
     public function testFindMostSuitable()
     {
         $scope = new Scope();
-        $provider = $this->createMock(ScopeCriteriaProviderInterface::class);
-        $provider->expects($this->once())
-            ->method('getCriteriaForCurrentScope')
-            ->willReturn(['fieldName' => 1]);
+        $fieldValue = new \stdClass();
+        $scopeCriteria = new ScopeCriteria(
+            ['fieldName' => $fieldValue, 'fieldName2' => null],
+            $this->scopeClassMetadata
+        );
 
-        $fieldsInfo = [['name' => 'fieldName'], ['name' => 'fieldName2']];
-        $scopeCriteria = new ScopeCriteria(['fieldName' => 1, 'fieldName2' => null], $fieldsInfo);
-        $repository = $this->getMockBuilder(ScopeRepository::class)->disableOriginalConstructor()->getMock();
-        $repository->expects($this->once())
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn(['fieldName', 'fieldName2']);
+
+        $provider = new StubScopeCriteriaProvider('fieldName', $fieldValue, \stdClass::class);
+
+        $this->scopeRepository->expects($this->once())
             ->method('findMostSuitable')
             ->with($scopeCriteria)
             ->willReturn($scope);
 
-        $this->entityStorage->expects($this->any())
-            ->method('getRepository')
-            ->willReturn($repository);
-
-        $this->entityFieldProvider->expects($this->once())
-            ->method('getRelations')
-            ->willReturn([
-                ['name' => 'fieldName'],
-                ['name' => 'fieldName2'],
-            ]);
-
-        $this->manager->addProvider('testScope', $provider);
-        $actualScope = $this->manager->findMostSuitable('testScope');
-        $this->assertEquals($scope, $actualScope);
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals($scope, $manager->findMostSuitable('testScope'));
     }
 
     /**
@@ -455,15 +576,21 @@ class ScopeManagerTest extends \PHPUnit\Framework\TestCase
      */
     public function testIsScopeMatchCriteria($expectedResult, $criteriaContext, $scopeFieldValue)
     {
-        $criteria = new ScopeCriteria($criteriaContext, []);
-
-        $this->manager->addProvider('some_type', new StubScopeCriteriaProvider());
         $scope = new StubScope();
         $scope->setScopeField($scopeFieldValue);
-        $this->entityFieldProvider->method('getRelations')->willReturn([]);
+        $scopeCriteria = new ScopeCriteria($criteriaContext, $this->scopeClassMetadata);
 
-        $result = $this->manager->isScopeMatchCriteria($scope, $criteria, 'some_type');
-        $this->assertEquals($expectedResult, $result);
+        $this->scopeClassMetadata->expects($this->once())
+            ->method('getAssociationNames')
+            ->willReturn([]);
+
+        $provider = new StubScopeCriteriaProvider('scopeField', new \stdClass(), \stdClass::class);
+
+        $manager = $this->getScopeManager(['testScope' => [$provider]]);
+        $this->assertEquals(
+            $expectedResult,
+            $manager->isScopeMatchCriteria($scope, $scopeCriteria, 'testScope')
+        );
     }
 
     /**
@@ -472,26 +599,26 @@ class ScopeManagerTest extends \PHPUnit\Framework\TestCase
     public function isScopeMatchCriteriaDataProvider()
     {
         return [
-            'scope match criteria' => [
-                'expectedResult' => true,
-                'criteriaContext' => [StubScopeCriteriaProvider::STUB_FIELD => 'expected_value'],
+            'scope match criteria'                             => [
+                'expectedResult'  => true,
+                'criteriaContext' => ['scopeField' => 'expected_value'],
                 'scopeFieldValue' => 'expected_value'
             ],
-            'scope dont match criteria' => [
-                'expectedResult' => false,
-                'criteriaContext' => [StubScopeCriteriaProvider::STUB_FIELD => 'unexpected_value'],
+            'scope dont match criteria'                        => [
+                'expectedResult'  => false,
+                'criteriaContext' => ['scopeField' => 'unexpected_value'],
                 'scopeFieldValue' => 'expected_value'
             ],
-            'scope match criteria with null value' => [
-                'expectedResult' => true,
-                'criteriaContext' => [StubScopeCriteriaProvider::STUB_FIELD => 'unexpected_value'],
+            'scope match criteria with null value'             => [
+                'expectedResult'  => true,
+                'criteriaContext' => ['scopeField' => 'unexpected_value'],
                 'scopeFieldValue' => null
             ],
             'scope dont match criteria with different objects' => [
-                'expectedResult' => false,
-                'criteriaContext' => [StubScopeCriteriaProvider::STUB_FIELD => new \stdClass()],
+                'expectedResult'  => false,
+                'criteriaContext' => ['scopeField' => new \stdClass()],
                 'scopeFieldValue' => new \stdClass()
-            ],
+            ]
         ];
     }
 }